feat(JOE-941,JOE-923,JOE-977): close milestone residuals
- Durable OpenCode: construct @opencode-ai/sdk/v2; session façade prefers
  v2.session.* with classic fallback for mocks (JOE-941)
- Channel residual: close dual-stack body migrate as security-shared +
  protocol freeze retained (JOE-923 scoped)
- Wiki: extract doctor deploy diagnostics leaf (JOE-977 progressive)
- Gate script updated for V2 construction; plan/matrix/burndown refreshed
