+++ products/gateway/src/channel-sync.ts
@@ -2,7 +2,7 @@ import * as fs from 'node:fs'
 import * as path from 'node:path'
 import { createHash, randomUUID } from 'node:crypto'
 import { DatabaseSync } from 'node:sqlite'
-import type { OpencodeClient } from '@opencode-ai/sdk'
+import type { OpencodeClient } from '@opencode-ai/sdk/v2'
 import type { ChannelAdapter } from './channels/provider.js'
 import { getConfig, getConfigDir } from './config.js'
 import { canCurrentDaemonWrite } from './daemon-leadership.js'
