+++ products/gateway/src/gateway-runtime.ts
@@ -1,4 +1,4 @@
-import type { OpencodeClient } from '@opencode-ai/sdk'
+import type { OpencodeClient } from '@opencode-ai/sdk/v2'
 
 /**
  * Typed process-wide holder for the daemon's live OpenCode client.
