+++ products/gateway/src/heartbeat.ts
@@ -5,7 +5,7 @@ import { schedulerCycleSnapshots } from './scheduler.js'
 import { createLogger } from './logger.js'
 import type { RunRecord, WorkState } from './work-store.js'
 import { getDaemonClient } from './gateway-runtime.js'
-import type { OpencodeClient } from '@opencode-ai/sdk'
+import type { OpencodeClient } from '@opencode-ai/sdk/v2'
 
 const log = createLogger({ component: 'heartbeat' })
