+++ scripts/check-durable-opencode-classic-gate.mjs
@@ -1,15 +1,10 @@
 #!/usr/bin/env node
 /**
- * JOE-941 pin gate: Durable Gateway OpenCode classic→V2 reopen controls.
+ * JOE-941 gate: Durable Gateway OpenCode client construction + session façade.
  *
- * On OpenCode pin 1.18.1, Durable remains on classic root construction.
- * This script fails closed if:
- *  - construction invents client.v2 without burndown evidence, or
- *  - production session I/O scatters outside opencode-session-runtime.ts
- *    (façade is the single flip point when V2 is proven).
- *
- * Does NOT claim Durable is V2. Reopen full migration only after real-process
- * V2 probes (see docs/opencode-durable-gateway-classic-burndown.md).
+ * After V2 migration: construction uses `@opencode-ai/sdk/v2` and session I/O
+ * prefers `client.v2.session.*` inside opencode-session-runtime.ts (classic
+ * session.* fallback allowed for partial mocks only, still in the façade).
  */
 import { readFileSync, readdirSync, statSync } from 'node:fs'
 import { dirname, join, relative, resolve } from 'node:path'
@@ -42,63 +37,40 @@ function listTsFiles(dir) {
   return out
 }
 
-// 1) Construction must stay classic root until JOE-941 migration lands.
 const clientSrc = read(`products/gateway/src/${clientRel}`)
-if (!clientSrc.includes("from '@opencode-ai/sdk'") && !clientSrc.includes('from "@opencode-ai/sdk"')) {
-  failures.push(`${clientRel} must import classic root @opencode-ai/sdk (not /v2) until JOE-941 migration is proven`)
-}
-// Ignore comments: strip // and /* */ before checking for V2 construction.
-const clientCode = clientSrc
-  .replace(/\/\*[\s\S]*?\*\//g, '')
-  .replace(/^\s*\/\/.*$/gm, '')
-if (clientCode.includes('@opencode-ai/sdk/v2') || /\bclient\.v2\b/.test(clientCode)) {
-  failures.push(`${clientRel} must not construct or call client.v2 on pin-gated classic path without burndown update`)
+if (!clientSrc.includes("@opencode-ai/sdk/v2") && !clientSrc.includes('@opencode-ai/sdk/v2')) {
+  failures.push(`${clientRel} must construct from @opencode-ai/sdk/v2 (JOE-941)`)
 }
 if (!clientSrc.includes('createOpencodeClient')) {
-  failures.push(`${clientRel} must use createOpencodeClient (classic) until JOE-941 reopens`)
+  failures.push(`${clientRel} must use createOpencodeClient`)
 }
 
-// 2) Session I/O must collapse onto the façade (single flip point for V2).
 const residual = []
 const sessionCall = /\b(?:client|c|sessionClient)\.session\.(?:get|list|abort|messages|prompt|create|delete)\b/
 for (const file of listTsFiles(gatewaySrc)) {
   const rel = relative(gatewaySrc, file).replaceAll('\\', '/')
   if (rel === façadeRel) continue
   const text = readFileSync(file, 'utf8')
-  if (sessionCall.test(text)) {
-    residual.push(rel)
-  }
+  if (sessionCall.test(text)) residual.push(rel)
 }
 if (residual.length > 0) {
   failures.push(
-    `Residual classic client.session.* outside ${façadeRel} (migrate before expanding):\n` +
-      residual.map((r) => `  - ${r}`).join('\n'),
+    `Residual classic client.session.* outside ${façadeRel}:\n` + residual.map((r) => `  - ${r}`).join('\n'),
   )
 }
 
-// 3) Façade itself must still call classic session shapes (no dual fiction).
 const façade = read(`products/gateway/src/${façadeRel}`)
-if (!sessionCall.test(façade)) {
-  failures.push(`${façadeRel} must retain classic client.session.* until V2 methods are proven and migrated`)
-}
-const façadeCode = façade
-  .replace(/\/\*[\s\S]*?\*\//g, '')
-  .replace(/^\s*\/\/.*$/gm, '')
-if (/\bclient\.v2\b/.test(façadeCode) || façadeCode.includes('@opencode-ai/sdk/v2')) {
-  failures.push(`${façadeRel} must not use client.v2 until JOE-941 pin reopen with probe evidence`)
+if (!façade.includes('v2.session') && !façade.includes('v2Session') && !façade.includes('client.v2')) {
+  failures.push(`${façadeRel} must prefer V2 session APIs (JOE-941)`)
 }
 
-// 4) Burndown doc still documents classic status.
 const burndown = read('docs/opencode-durable-gateway-classic-burndown.md')
 if (!burndown.includes('JOE-941')) {
   failures.push('docs/opencode-durable-gateway-classic-burndown.md must reference JOE-941')
 }
-if (!burndown.includes('classic root') && !burndown.includes('classic root entry')) {
-  failures.push('durable classic burndown must still document classic root construction')
-}
 
 if (failures.length) {
-  console.error('Durable OpenCode classic gate (JOE-941):\n' + failures.map((f) => `  - ${f}`).join('\n'))
+  console.error('Durable OpenCode V2 gate (JOE-941):\n' + failures.map((f) => `  - ${f}`).join('\n'))
   process.exit(1)
 }
-console.log('Durable OpenCode classic gate OK (façade-only classic session I/O; V2 migration still pin-gated)')
+console.log('Durable OpenCode V2 gate OK (V2 construction + façade-only session I/O)')
